projects/academic_staff_management_system/backend/requirements.txt
django
djangorestframework
gunicorn
Pillow
djangorestframework-simplejwt
django-filter